+++ luscii_patient_actions_sdk_android/android/src/main/kotlin/nl/digizorg/errors/LusciiFlutterSdkError.kt
@@ -6,4 +6,5 @@ enum class LusciiFlutterSdkError(val code: String, val message: String) {
     INVALID_API_KEY("2", "Invalid API Key"),
     UNAUTHORIZED("3", "Unauthorized"),
     NOT_INITIALIZED("4", "Luscii SDK not initialized"),
+    UNSUPPORTED_VERSION("5", "Luscii SDK requires iOS 17.0 or higher"), // Only used by iOS
 }